requirements.txt
torch==2.2.2
transformers==4.45.2
datasets==2.20.0
evaluate==0.4.2
accelerate==0.34.2
scikit-learn==1.5.2
pyyaml==6.0.2
tqdm==4.66.5
numpy==1.26.4
